[Patch v6.9.39] split strategy module
